move leftView to hFieldTextContent
